Add Closeable
