+++ src/main/java/com/spotify/confidence/eventsender/EventSenderEngine.java
@@ -1,9 +1,9 @@
 package com.spotify.confidence.eventsender;
 
-interface EventSenderEngine {
+import java.io.Closeable;
+
+interface EventSenderEngine extends Closeable {
   void send(String name, ConfidenceValue.Struct message, ConfidenceValue.Struct context);
 
   void send(String name, ConfidenceValue.Struct context);
-
-  void shutdown();
 }
